+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [0.8.6]
+- Experimental fix to avoid inifinte animation loop caused by multi-level NotificationListener (thanks @rienkkk)
+
+## [0.8.5]
+- Experimental fix for infinite loop bug on small listview area (thanks @rienkkk)
+- Expose listview padding to param (thanks @siloebb)
+
+## [0.8.4]
+- Experimental fix overflow when changing orientation (thanks @mirrorlink)
+
+## [0.8.3]
+- Add flag to control to dispatch scroll notifications to further ancestors (thanks @msangals)
+
 ## [0.8.2]
 - Experimental fix for iOS over-scroll bug